Revert "Fix bug with Dashbord Repository"
This reverts commit f11798440f2947ed0fe7f5bb990d4c4f2dae05d4.

+++ CharlieBackend.Core/DTO/Dashboard/StudentGroupResults/AverageStudentGroupMarkDto.cs
@@ -10,6 +10,6 @@ namespace CharlieBackend.Core.DTO.Dashboard
 
         public long StudentGroupId { get; set; }
 
-        public decimal? AverageMark { get; set; }
+        public double? AverageMark { get; set; }
     }
 }
